requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
plotly>=5.18.0
numpy>=1.24.0
openpyxl>=3.0.0
ultralytics>=8.0.0
torch
opencv-python
transformers
pillow
